requirements.txt
openai~=2.30.0
python-dotenv~=1.2.2
python-telegram-bot~=22.7
pandas~=3.0.1